requirements.txt
certifi == 2024.2.2
chardet==5.2.0
geopy==2.4.1
idna==3.6
pyTelegramBotAPI==4.16.1
requests==2.31.0
urllib3==2.2.1
python-dotenv == 1.0.1